chore(profile): organizing profile domain
