+++ src-tauri/src/profiles/mod.rs
@@ -1,3 +1,5 @@
+pub mod actions;
+
 use serde::{Deserialize, Serialize};
 use std::collections::HashMap;
 use std::fmt::Display;
